+++ Libraries/SideScroll.Avalonia/Extensions/DataGridExtensions.cs
@@ -170,40 +170,6 @@ public static class DataGridExtensions
 		return TableToCsv(columns, contentRows);
 	}
 
-	private static string TableToCsv(List<ColumnInfo> columns, List<List<string>> contentRows)
-	{
-		StringBuilder stringBuilder = new();
-		bool addComma = false;
-		foreach (ColumnInfo columnInfo in columns)
-		{
-			if (addComma)
-				stringBuilder.Append(',');
-			addComma = true;
-			stringBuilder.Append(columnInfo.Name);
-		}
-		stringBuilder.Append('\n');
-
-		foreach (var row in contentRows)
-		{
-			addComma = false;
-			foreach (string value in row)
-			{
-				if (addComma)
-					stringBuilder.Append(',');
-				addComma = true;
-
-				string text = value ?? "";
-				text = text.Replace("\"", "\"\""); // escape double quote
-				stringBuilder.Append('"');
-				stringBuilder.Append(text);
-				stringBuilder.Append('"');
-			}
-			stringBuilder.Append('\n');
-		}
-
-		return stringBuilder.ToString();
-	}
-
 	private static void GetDataGridContents(DataGrid dataGrid, IEnumerable items, out List<ColumnInfo> columns, out List<List<string>> contentRows, int? maxValueLength = null)
 	{
 		columns = [];